Finish footer layout and behavior
