+++ src/assets/styles/GlobalStyle.js
@@ -12,7 +12,7 @@ const GlobalStyle = createGlobalStyle`
     h3,
     h4,
     h5,
-    h6,
+    h3,
     p,
     blockquote,
     pre,
