+++ src/services/api.js
@@ -1,17 +1,20 @@
+/*
 import axios from "axios";
 
 const API_URL = "";
 
 const createHeaders = (token) => {
-    return { 
-        headers: { 
-            Authorization: `Bearer ${token}` 
-        } 
+    return {
+        headers: {
+            Authorization: `Bearer ${token}`
+        }
     };
 }
 
-//const sendNewRecord = ( token, body ) => axios.post(`${API_URL}/records`, body, createHeaders(token));
+const sendNewRecord = ( token, body ) =>
+    axios.post(`${API_URL}/records`, body, createHeaders(token));
 
 export {
 
 }
+*/
